+++ backend/src/index.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import cors from "cors";
 import "dotenv/config";
-import mongoose, { ConnectOptions } from "mongoose";
+import mongoose from "mongoose";
 import cookieParser from "cookie-parser";
 
 import userRoutes from "./routes/users";
@@ -14,12 +14,12 @@ import uploadRoutes from "./routes/upload";
 import downloadRoutes from "./routes/download";
 import notificationRoutes from "./routes/notification";
 import notificationSettingRoutes from "./routes/notification-settings";
+import supportTicketRoutes from "./routes/support-ticket";
 import configRoutes from "./system/config";
 
 import logger from "./utils/logger";
 import swaggerDocs from "./utils/swagger";
 import { startChangeStream } from "./utils/change-stream";
-import path from "path";
 import { auditAssets } from "./utils/common";
 import {
   autoMail,
@@ -34,29 +34,15 @@ import {
 import autoMailRoutess from "./system/automail";
 import backupRoutes, { listCollection } from "./system/backup";
 import versionRoutes from "./system/version";
+import path from "path";
 
 const DEFAULT_PORT = 3000;
 const port = Number(process.env.PORT) || DEFAULT_PORT;
 
 const HOST = (process.env.HOST as string) || "localhost";
-const FRONTENDLOC = "../../frontend/dist";
-const NODE_ENV = process.env.NODE_ENV || "development";
+const FRONTEND_BUILD_LOCATION = "../../frontend/dist";
 
-const connectOptions = {
-  development: {},
-  staging: {
-    authSource: process.env.MONGODB_AUTH_SOURCE,
-    user: process.env.MONGODB_USERNAME,
-    pass: process.env.MONGODB_PASSWORD,
-    replicaSet: "rs0",
-  },
-  production: {},
-}[NODE_ENV] as ConnectOptions;
-
-mongoose.connect(process.env.MONGODB_CONNECTION_STRING as string, {
-  dbName: process.env.MONGODB_NAME,
-  ...connectOptions,
-});
+mongoose.connect(process.env.MONGODB_CONNECTION_STRING as string);
 const db = mongoose.connection;
 
 db.on("connected", () => {
@@ -71,6 +57,7 @@ startChangeStream();
 
 const app = express();
 
+app.use(express.static(path.join(__dirname, FRONTEND_BUILD_LOCATION)));
 app.use(cookieParser());
 app.use(express.json({ limit: "200mb" }));
 app.use(express.urlencoded({ limit: "200mb", extended: true }));
@@ -80,8 +67,6 @@ app.use(
     credentials: true,
   })
 );
-
-app.use(express.static(path.join(__dirname, FRONTENDLOC)));
 
 app.use("/api/auth", authRoutes);
 app.use("/api/users", userRoutes);
@@ -94,16 +79,20 @@ app.use("/api/download", downloadRoutes);
 app.use("/api/download", downloadRoutes);
 app.use("/api/notification", notificationRoutes);
 app.use("/api/notification_settings", notificationSettingRoutes);
+app.use("/api/support_ticket", supportTicketRoutes);
 app.use("/config", configRoutes);
 app.use("/autoMail", autoMailRoutess);
 app.use("/backup", backupRoutes);
 app.use("/version", versionRoutes);
+app.use("/healthcheck", (_, res) =>
+  res.send("Inventory backend is running...")
+);
 
-// Catch-all route for unmatched URLs (place it here)
+// Catch-all route for unmatched URLs
 if (process.env.NODE_ENV !== "development") {
   app.get("/*", function (req, res) {
     res.sendFile(
-      path.join(__dirname, `${FRONTENDLOC}/index.html`),
+      path.join(__dirname, `${FRONTEND_BUILD_LOCATION}/index.html`),
       function (err) {
         if (err) {
           res.status(500).send(err);