+++ frontend/src/components/setttings-ui/load-backup/LoadBackupForm.tsx
@@ -1,68 +1,75 @@
-import { useState, useEffect } from 'react';
-import { useForm } from 'react-hook-form';
+import { useState, useEffect } from "react";
+import { useForm } from "react-hook-form";
 import {
   Form,
   FormField,
   FormItem,
   FormControl,
-  FormDescription
-} from '@/components/ui/form';
-import { Separator } from '@/components/ui/separator';
-import { FileUploader } from '@/components/ui/file-uploader';
-import { BackupValidationModal } from './BackupValidationModal';
-import { validateBackupFile } from '@/ims-service';
-import { ValidationResult } from '@/types/backup';
-
+  FormDescription,
+} from "@/components/ui/form";
+import { Separator } from "@/components/ui/separator";
+import { FileUploader } from "@/components/ui/file-uploader";
+import { BackupValidationModal } from "./BackupValidationModal";
+import { validateBackupFile } from "@/ims-service";
+import { ValidationResult } from "@/types/backup";
+import { Accept } from "react-dropzone";
 
 const SystemBackup = () => {
   const form = useForm({
     defaultValues: {
-      file: ''
-    }
+      file: "",
+    },
   });
 
   const [uploadedFile, setUploadedFile] = useState<File | undefined>();
-  const [, setFileAsBase64] = useState<{ src: string }>({ src: '' });
-  const [validationResult, setValidationResult] = useState<ValidationResult>({ message: '' });
-  const [validationComplete, setValidationComplete] = useState<boolean | null>(null);
+  const [, setFileAsBase64] = useState<{ src: string }>({ src: "" });
+  const [validationResult, setValidationResult] = useState<ValidationResult>({
+    message: "",
+  });
+  const [validationComplete, setValidationComplete] = useState<boolean | null>(
+    null
+  );
 
-  const acceptedFileTypes: string = '.zip,application/octet-stream,application/zip,application/x-zip,application/x-zip-compressed'
+  const acceptedFileTypes: Accept = {
+    "application/octet-stream": [],
+    "application/zip": [".zip"],
+    "application/x-zip": [],
+    "application/x-zip-compressed": [],
+  };
 
-  const handleFile = (file: File | undefined) => {
+  const handleFile = (files: File[] | undefined) => {
+    const file = files?.[0];
     setUploadedFile(file);
-    
+
     if (file !== undefined) {
-      setValidationResult({message: ''})
+      setValidationResult({ message: "" });
       const reader = new FileReader();
       reader.onloadend = async () => {
         const base64String = reader.result as string;
-        const plainText = base64String.split(',')[1]
-        const backendExpected = "data:application/zip;base64," + plainText
+        const plainText = base64String.split(",")[1];
+        const backendExpected = "data:application/zip;base64," + plainText;
         const payload = {
           src: backendExpected,
         };
         setFileAsBase64(payload);
-        try {
-          setValidationComplete(false)
-          const validation: ValidationResult = await validateBackupFile(payload);
-          setValidationResult(validation);
-          setValidationComplete(true);
-        } catch (err) {
-          throw err;
-        }
+
+        setValidationComplete(false);
+        const validation: ValidationResult = await validateBackupFile(payload);
+        setValidationResult(validation);
+        setValidationComplete(true);
       };
       reader.readAsDataURL(file);
     } else {
-      setFileAsBase64({ src: '' })
+      setFileAsBase64({ src: "" });
     }
-  }
+  };
 
   useEffect(() => {
     if (!uploadedFile) {
-      setValidationResult({ message: '' });
+      setValidationResult({ message: "" });
       setValidationComplete(null);
-    };
-  }, [uploadedFile])
+    }
+  }, [uploadedFile]);
 
   return (
     <div className="flex flex-col md:w-5/6 max-w-4xl">
@@ -73,9 +80,7 @@ const SystemBackup = () => {
         </h3>
       </div>
       <Form {...form}>
-        <form
-          className="w-full flex flex-col"
-        >
+        <form className="w-full flex flex-col">
           <FormField
             control={form.control}
             name="file"
@@ -83,8 +88,8 @@ const SystemBackup = () => {
               <FormItem className="pb-4">
                 <FormControl className="pb-4">
                   <FileUploader
-                    handleFile={handleFile}
-                    uploadedFile={uploadedFile}
+                    handleFiles={handleFile}
+                    uploadedFiles={uploadedFile ? [uploadedFile] : []}
                     accept={acceptedFileTypes}
                     {...field}
                   />
@@ -97,12 +102,18 @@ const SystemBackup = () => {
           />
           <Separator className="my-4" />
           <div className="flex flex-row justify-end gap-4">
-            <BackupValidationModal result={validationResult} validationComplete={validationComplete} onComplete={() => {setUploadedFile(undefined)}} />
+            <BackupValidationModal
+              result={validationResult}
+              validationComplete={validationComplete}
+              onComplete={() => {
+                setUploadedFile(undefined);
+              }}
+            />
           </div>
         </form>
       </Form>
     </div>
-  )
-}
+  );
+};
 
 export default SystemBackup;